History component to now retrieve full & accurate information unconditionally
