+++ src/Components/Navbar.scss
@@ -34,6 +34,10 @@
       &:hover {
         background-color: rgb(166, 231, 228);
       }
+
+      &.no-hover:hover {
+        background-color: white;
+      }
     }
 
     .active-item {
