+++ src/Components/NewTranscript.scss
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   background-color: rgb(250, 250, 250);
-  height: 100%;
+  min-height: 100%;
 
   .controls {
     user-select: none;
@@ -129,7 +129,12 @@
     padding: 1rem 0.5rem;
     height: 480px;
     margin: 1rem 2.5rem;
-    border-radius: 16px;
+    border-radius: 12px;
+    resize: none;
+    width: 100%;
+    line-height: 1.5;
+    max-width: 1440px;
+    align-self: center;
   }
 
   .text, .title {
@@ -139,6 +144,7 @@
 
     &.read-only {
       background-color: rgb(241, 241, 241);
+      pointer-events: none;
     }
   }
